<script>s are not necessarily JavaScript.
We use jQuery templates, which expects <script
type="text/x-jquery-tmpl"> tags for the template contents. (There are
reasons to use <script> instead of e.g. <div>, but those aren't really
relevant; the point is that <script> is not necessarily JS.)

This patch checks the language processors to ensure the called type
exists before trying to call a non-existent method (the code was here to
support this, but the method call wasn't properly guarded.)

I also modified the browser tests to check for this.