modified:   src/zombie/resources.coffee
According RFC2616 4.4 Message Length:

" 3.If a Content-Length header field (section 14.13) is present, its
decimal value in OCTETs represents both the entity-length and the
transfer-length. The Content-Length header field MUST NOT be
sent if these two lengths are different (i.e., if a Transfer-Encoding
header field is present). If a message is received with
both a Transfer-Encoding header field and a Content-Length header field,
the latter MUST be ignored"

So I change code not to send headers['content-length'],
if a form is "multipart/form-data".

This change also make zombie work well with file uploading libraries,
like formidable and parted.