Treat curly rule violation as warning, not error.
There are a few tests that use `for` loops with a single statement,
and this rule wants curly braces to be removed from such for loops.

I'm not sure what the code standard for this project is, so I went
for the simplest way to make the build pass. (On node 5.3.0).