Clear cookie header on redirect
In the event that a redirect invalidates the last cookie, the
cookie store is cleared but the cookie header from the previous
request remains. Since the request pipeline recomputes the cookie
header for each request, this problem can be addressed by
removing the cookie header from the request prior to redirecting.