More complete documentation of Browser module API
Many of the Browser methods present in Zombie are missing from the README which makes it hard to use as a new user. The descriptions are pulled from `src/dom/index.js`. I'm unsure if this is the best way to organize the API docs but now many more relevant methods are in the README.